third_party/blink/web_tests/fast/table/table-rowspan-cell-override-logical-content-height-reset-issue.html

<!DOCTYPE html>
<style>
    td, textarea { font: 20px/1 Ahem }
</style>
<script src="../../resources/check-layout.js"></script>
<body>
    <h3>Test for chromium bug : <a href="https://code.google.com/p/chromium/issues/detail?id=445253">445253</a>. Setting rowspan=2 on the table cell causes the whole table to resize again and again.</h3>
    <table border="1" data-expected-height="60">
        <tr>
            <td rowspan="2" id="ta"><textarea style='height:100%;'>Text area text</textarea></td>
        </tr>
    </table>
</body>

<script>
    document.body.offsetTop;
    document.getElementById("ta").rowSpan="2";
    checkLayout('table');
</script>
