third_party/blink/web_tests/flag-specific/disable-layout-ng/fast/table/table-rowspan-cell-override-logical-content-height-reset-issue-expected.txt

Test for chromium bug : 445253. Setting rowspan=2 on the table cell causes the whole table to resize again and again.
FAIL:
Expected 60 for height, but got 58. 

<table border="1" data-expected-height="60">
        <tbody><tr>
            <td rowspan="2" id="ta"><textarea style="height:100%;">Text area text</textarea></td>
        </tr>
    </tbody></table>
